+++ packages/web3-eth-iban/test/unit/iban.test.ts
@@ -1,4 +1,4 @@
-import { Iban } from '../../src/iban';
+import { Iban } from '../../src/index';
 import {
 	validIbanToAddressData,
 	validFromBbanData,
